app/services/zodiac.py
from datetime import date
from typing import Optional
import logging

logger = logging.getLogger(__name__)

# Date ranges inclusive for western sun-sign
ZODIAC_RANGES = [
    ("Capricorn", (12, 22), (1, 19)),
    ("Aquarius", (1, 20), (2, 18)),
    ("Pisces", (2, 19), (3, 20)),
    ("Aries", (3, 21), (4, 19)),
    ("Taurus", (4, 20), (5, 20)),
    ("Gemini", (5, 21), (6, 20)),
    ("Cancer", (6, 21), (7, 22)),
    ("Leo", (7, 23), (8, 22)),
    ("Virgo", (8, 23), (9, 22)),
    ("Libra", (9, 23), (10, 22)),
    ("Scorpio", (10, 23), (11, 21)),
    ("Sagittarius", (11, 22), (12, 21)),
]

def get_zodiac_sign(birth_date: date) -> str:
    """
    Get zodiac sign from birth date using tropical zodiac
    
    Args:
        birth_date: Date of birth
        
    Returns:
        Zodiac sign name
    """
    m, d = birth_date.month, birth_date.day
    
    for name, (m1, d1), (m2, d2) in ZODIAC_RANGES:
        # if range doesn't cross year (most):
        if m1 < m2:
            if (m == m1 and d >= d1) or (m == m2 and d <= d2) or (m1 < m < m2):
                return name
        else:
            # crosses year boundary (Capricorn)
            if (m == m1 and d >= d1) or (m == m2 and d <= d2) or (m > m1 or m < m2):
                return name
    
    return "Capricorn"


def get_ascendant(birth_date: date, birth_time: str, latitude: float, longitude: float) -> Optional[str]:
    """
    Calculate ascendant (Lagna) from birth details
    This is a stub for Panchang integration
    
    In production, this would use libraries like:
    - swisseph (Swiss Ephemeris)
    - jyotisha (Python astrological library)
    - pyephem (Astronomical calculations)
    
    Example implementation:
    ```python
    import swisseph as swe
    from datetime import datetime
    
    # Convert birth date and time to Julian day
    dt = datetime.combine(birth_date, parse_time(birth_time))
    jd = swe.julday(dt.year, dt.month, dt.day, 
                    dt.hour + dt.minute/60.0, swe.GREG_CAL)
    
    # Calculate ascendant
    cusps, ascmc = swe.houses(jd, latitude, longitude, b'P')
    ascendant_deg = ascmc[0]
    
    # Convert to zodiac sign
    sign_num = int(ascendant_deg / 30)
    signs = ["Aries", "Taurus", "Gemini", "Cancer", "Leo", "Virgo",
             "Libra", "Scorpio", "Sagittarius", "Capricorn", "Aquarius", "Pisces"]
    return signs[sign_num]
    ```
    
    Args:
        birth_date: Date of birth
        birth_time: Time of birth (HH:MM format)
        latitude: Latitude of birth place
        longitude: Longitude of birth place
        
    Returns:
        Ascendant sign name or None if calculation fails
    """
    logger.info("Ascendant calculation stub - integrate with Panchang library for production")
    # Stub: Return None for now
    # In production, integrate with swisseph or jyotisha
    return None


def get_moon_sign(birth_date: date, birth_time: str, latitude: float, longitude: float) -> Optional[str]:
    """
    Calculate moon sign (Rashi) from birth details
    This is a stub for Panchang integration
    
    Similar to get_ascendant, this would use astronomical calculations
    to determine the moon's position at birth time.
    
    Args:
        birth_date: Date of birth
        birth_time: Time of birth (HH:MM format)
        latitude: Latitude of birth place
        longitude: Longitude of birth place
        
    Returns:
        Moon sign name or None if calculation fails
    """
    logger.info("Moon sign calculation stub - integrate with Panchang library for production")
    # Stub: Return None for now
    # In production, integrate with swisseph or jyotisha
    return None


def get_panchang_data(birth_date: date, birth_time: str, latitude: float, longitude: float) -> dict:
    """
    Get comprehensive Panchang data (Tithi, Nakshatra, Yoga, Karana, etc.)
    This is a stub for full Panchang integration
    
    Args:
        birth_date: Date of birth
        birth_time: Time of birth
        latitude: Latitude of birth place
        longitude: Longitude of birth place
        
    Returns:
        Dictionary with Panchang data (stub returns empty dict)
    """
    logger.info("Panchang data calculation stub - integrate with Panchang API/library for production")
    return {
        "tithi": None,
        "nakshatra": None,
        "yoga": None,
        "karana": None,
        "ascendant": get_ascendant(birth_date, birth_time, latitude, longitude),
        "moon_sign": get_moon_sign(birth_date, birth_time, latitude, longitude)
    }
