requirements.txt
fastapi
uvicorn
pydantic
cohere
python-dotenv
numpy
faiss-cpu
pytest
pytest-asyncio
httpx
